% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1capstoneproject2.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{cleans the LOCATION_NAME column by stripping out the country name (including the colon) and
converts names to title case (as opposed to all caps)}
\usage{
eq_clean_data(mydf)
}
\arguments{
\item{mydf}{that contains location names written in upper case}
}
\value{
a dataframe filtered required for mapping in a timeline the data
}
\description{
cleans the LOCATION_NAME column by stripping out the country name (including the colon) and
converts names to title case (as opposed to all caps)
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes-2021-02-17_20-53-07_-0500.tsv",package="mycapstone") ##XX
eq_clean_data(eq_data_read(filename))
}

}
