% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1capstoneproject2.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates pop up text for markers.
generates HTML formatted text for popups for map markers.}
\usage{
eq_create_label(eq_clean = NULL)
}
\arguments{
\item{eq_clean}{The clean earthquake data in a tbl_df object.}
}
\value{
returns a character vector containing popup text to be used in a leaflet visualization.
}
\description{
Creates pop up text for markers.
generates HTML formatted text for popups for map markers.
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes_data.txt.zip",package="capstone")
eq_location_clean(eq_clean_data(eq_data_read(filename))) \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(datetime) >= 1980) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "popup_text")
}

}
