% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week1.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleaning the location
This function creates a column including the country where the earthquake has
happened in title case format.}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{data.frame -- the earthquake data base by NOAA (USA)}
}
\value{
data.frame
}
\description{
Cleaning the location
This function creates a column including the country where the earthquake has
happened in title case format.
}
\examples{
\dontrun{eq_location_clean(data)}
}
