% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1capstoneproject2.R
\name{eq_map}
\alias{eq_map}
\title{Earthquakes Data in an Interactive Map.}
\usage{
eq_map(eq_clean = NULL, annot_col = "datetime")
}
\arguments{
\item{eq_clean}{The clean earthquake data in a tbl_df object.}

\item{annot_col}{Column in the tbl_df object to be used for annotation.}
}
\value{
returns an interactive map.
}
\description{
mapped centered with their latitude and longitude "epicenter" which is annotated based on an annot_col which the user can specify.
plus specifies "popup_text" by a call to eq_create_label generates the appropriate text..
}
\note{
warning an invalid column name & uses the LOCATION_NAME column as annotation column.
}
\examples{
\dontrun{
filename<-system.file("data","earthquakes-2021-02-17_20-53-07_-0500.tsv",package="mycapstone")
eq_location_clean(eq_clean_data(eq_data_read(filename))) \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(datetime) >= 1980) \%>\%
eq_map(annot_col = "datetime")
}

}
\references{
\url{http://rstudio.github.io/leaflet/}
}
