% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week2.R
\name{geom_timeline}
\alias{geom_timeline}
\title{This is my capstone project during the course "Mastering Software Development
in R Capstone" by Roger D. Peng and Brooke Anderson, Johns Hopkins University
as a part of "Mastering Software Development in R Professional Certificate"
program on Coursera.
-----------------------------------------------------------------------------
WEEK 2 - Visualize
-----------------------------------------------------------------------------
for Windows, if you get an error about Java
read data: Earthquake Strong Motion Data (1933-1994) by National Centers for
Environmenta Information, NOAA, USA
<https://www.ngdc.noaa.gov/hazard/strong.shtml>}
\usage{
geom_timeline(data, y_axis, color_vals, size_val, alpha_val)
}
\description{
This is my capstone project during the course "Mastering Software Development
in R Capstone" by Roger D. Peng and Brooke Anderson, Johns Hopkins University
as a part of "Mastering Software Development in R Professional Certificate"
program on Coursera.
-----------------------------------------------------------------------------
WEEK 2 - Visualize
-----------------------------------------------------------------------------
for Windows, if you get an error about Java
read data: Earthquake Strong Motion Data (1933-1994) by National Centers for
Environmenta Information, NOAA, USA
<https://www.ngdc.noaa.gov/hazard/strong.shtml>
}
