% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1capstoneproject2.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Ploting an Earthquake's Location timeline
building a GEOM Function from scratch
use a Dataframe compiled using the function eq_clean_data.
a prototype function as foundation for geom_timeline function.
use the ggplot2's geom_point.
use the Earthquakes' dates as X-axis main values
The geom_point's size and colour defined by the Earthquake's magnitude
The GeomTimeLine was build using the Function Prototype provided in the Course's Material 4.7.1 Building a New Geom}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
Ploting an Earthquake's Location timeline
building a GEOM Function from scratch
use a Dataframe compiled using the function eq_clean_data.
a prototype function as foundation for geom_timeline function.
use the ggplot2's geom_point.
use the Earthquakes' dates as X-axis main values
The geom_point's size and colour defined by the Earthquake's magnitude
The GeomTimeLine was build using the Function Prototype provided in the Course's Material 4.7.1 Building a New Geom
}
\keyword{datasets}
