% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week1.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{This is my capstone project during the course "Mastering Software Development
in R Capstone" by Roger D. Peng and Brooke Anderson, Johns Hopkins University
as a part of "Mastering Software Development in R Professional Certificate"
program on Coursera.
-----------------------------------------------------------------------------
WEEK 1 - Clean
-----------------------------------------------------------------------------
Clean earthquake data
This function cleans the earthquake data retrieved from NOAA database. It
uses the archived data including the records dated before 1995.}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{data.frame -- the earthquake data base by NOAA (USA)}
}
\value{
data.frame
}
\description{
This is my capstone project during the course "Mastering Software Development
in R Capstone" by Roger D. Peng and Brooke Anderson, Johns Hopkins University
as a part of "Mastering Software Development in R Professional Certificate"
program on Coursera.
-----------------------------------------------------------------------------
WEEK 1 - Clean
-----------------------------------------------------------------------------
Clean earthquake data
This function cleans the earthquake data retrieved from NOAA database. It
uses the archived data including the records dated before 1995.
}
\examples{
\dontrun{eq_clean_data(data)}
}
