% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week3.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creating html label
This function creates a html label including location and magnitude of each
earthquake and the number of deaths in them. The output can be used as a
popup message for \link{eq_map} function.}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
character
}
\description{
Creating html label
This function creates a html label including location and magnitude of each
earthquake and the number of deaths in them. The output can be used as a
popup message for \link{eq_map} function.
}
\examples{
\dontrun{
data <- eq_location_clean(eq_clean_data(data))
data <- dplyr::mutate(data, popup_text = eq_create_label(data))
eq_map(data, "popup_text")}
}
